requirements.txt
fastapi>=0.75.0
requests>=2.27.1
python-dotenv>=0.20.0
pyturf>=0.6.0
geographiclib>=1.52
numpy>=1.22.3
geopy>=2.2.0
basemap>=1.3.2
scikit-learn>=1.0.2
geojson-pydantic>=0.3.3